test(select): migrate compare-with
